+++ model/alert_message.go
@@ -69,47 +69,3 @@ func (n *AlertMessage) ConvertLevelToInt() string {
 
 	return AlertLevel[nLevel]
 }
-
-type Alerts []Alert
-
-// 定义告警级别的优先级映射
-// var levelPriority = map[string]int{
-// 	"信息":   0,
-// 	"警告":   1,
-// 	"一般严重": 2,
-// 	"严重":   3,
-// 	"灾难":   4,
-// }
-
-// 获取告警级别的优先级值，默认返回最小优先级
-func getLevelPriority(alert Alert) int {
-	levelStr, exists := alert.Labels["level"]
-	if !exists {
-		return 0 // 如果没有找到 level 字段，返回默认值
-	}
-
-	nLevel, err := strconv.Atoi(levelStr)
-	if err != nil || nLevel < 0 {
-		return 0 // 如果转换失败或数值无效，返回默认值
-	}
-
-	return nLevel
-}
-
-// 实现 sort.Interface 接口的方法
-func (a Alerts) Len() int {
-	return len(a)
-}
-
-func (a Alerts) Swap(i, j int) {
-	a[i], a[j] = a[j], a[i]
-}
-
-func (a Alerts) Less(i, j int) bool {
-	// 获取两个告警的优先级
-	priorityI := getLevelPriority(a[i])
-	priorityJ := getLevelPriority(a[j])
-
-	// 从大到小排序，所以这里是比较priorityJ是否小于priorityI
-	return priorityJ < priorityI
-}
